SQL SystemDB test updated to allow Spansh/EDSM to work

+++ TestSQL/SystemDB/SQLiteConnectionSystem.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright 2015-2021 EDDiscovery development team
+ * Copyright 2015-2023 EDDiscovery development team
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at
@@ -10,8 +10,6 @@
  * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
  * ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
- * 
- * EDDiscovery is not affiliated with Frontier Developments plc.
  */
  
 using SQLLiteExtensions;
@@ -29,17 +27,17 @@ namespace EliteDangerousCore.DB
 
         const string tablepostfix = "temp"; // postfix for temp tables
 
-        public bool UpgradeSystemsDB()
+        // will throw on error
+        // returns 0 DB was good, else version number
+        public int UpgradeSystemsDB()
         {
-            try
+            // BE VERY careful with connections when creating/deleting tables - you end up with SQL Schema errors or it not seeing the table
+
+            SQLExtRegister reg = new SQLExtRegister(this);
+            int dbver = reg.GetSetting("DBVer", (int)0);      // use reg, don't use the built in func as they create new connections and confuse the schema
+
+            if (dbver < 210)    // less than 210, delete the lot and start again
             {
-                ExecuteNonQuery("CREATE TABLE IF NOT EXISTS Register (ID TEXT PRIMARY KEY NOT NULL, ValueInt INTEGER, ValueDouble DOUBLE, ValueString TEXT, ValueBlob BLOB)");
-
-                // BE VERY careful with connections when creating/deleting tables - you end up with SQL Schema errors or it not seeing the table
-
-                SQLExtRegister reg = new SQLExtRegister(this);
-                int dbver = reg.GetSetting("DBVer", (int)0);      // use reg, don't use the built in func as they create new connections and confuse the schema
-
                 ExecuteNonQueries(new string[]                  // always set up
                     {
                     "DROP TABLE IF EXISTS Distances",
@@ -54,9 +52,9 @@ namespace EliteDangerousCore.DB
 
                 var tablesql = this.SQLMasterQuery("table");
                 int index = tablesql.FindIndex(x => x.TableName == "Systems");
-                bool oldsystems = index >=0 && tablesql[index].SQL.Contains("commandercreate");
+                bool oldsystems = index >= 0 && tablesql[index].SQL.Contains("commandercreate");
 
-                if ( dbver < 200 || oldsystems)
+                if (dbver < 200 || oldsystems)
                 {
                     ExecuteNonQueries(new string[]             // always kill these old tables 
                     {
@@ -64,7 +62,7 @@ namespace EliteDangerousCore.DB
                     });
                 }
 
-                if (dbver < 102)        // is it older than 102, its unusable
+                if (dbver < 200)                           // is it older than 200, now unusable.  Removed 102-200 reformat code as its been long enough
                 {
                     ExecuteNonQueries(new string[]         // older than 102, not supporting, remove
                     {
@@ -79,24 +77,23 @@ namespace EliteDangerousCore.DB
                 CreateSystemDBTableIndexes();           // ensure they are there 
                 DropStarTables(tablepostfix);         // clean out any temp tables half prepared 
 
-                //    conn.Vacuum();  // debug
-
-                if (dbver < 200)
-                {
-                    reg.PutSetting("DBVer", (int)200);
-                }
-
-                return true;
+                return 211;     // return, this makes the whole schebang up to 211
             }
-            catch (Exception ex)
+           
+            if ( dbver < 211)           // if we had 210, but not 211, we need to add a new column in but keep the data
             {
-                System.Windows.Forms.MessageBox.Show("UpgradeSystemsDB error: " + ex.Message + Environment.NewLine + ex.StackTrace);
-                return false;
+                ExecuteNonQueries(new string[]                  // always set up
+                    {
+                        "ALTER TABLE Systems ADD COLUMN info INT DEFAULT NULL",
+                    });
+
+                return 211;
             }
+
+            return 0;
         }
 
         #endregion
-
 
         #region Helpers
 
@@ -107,7 +104,7 @@ namespace EliteDangerousCore.DB
                 // purposely not using autoincrement or unique on primary keys - this slows it down.
 
                 "CREATE TABLE IF NOT EXISTS Sectors" + postfix + " (id INTEGER PRIMARY KEY NOT NULL, gridid INTEGER, name TEXT NOT NULL COLLATE NOCASE)",
-                "CREATE TABLE IF NOT EXISTS Systems" + postfix + " (edsmid INTEGER PRIMARY KEY NOT NULL , sectorid INTEGER, nameid INTEGER, x INTEGER, y INTEGER, z INTEGER)",
+                "CREATE TABLE IF NOT EXISTS Systems" + postfix + " (edsmid INTEGER PRIMARY KEY NOT NULL , sectorid INTEGER, nameid INTEGER, x INTEGER, y INTEGER, z INTEGER, info INTEGER DEFAULT NULL)",
                 "CREATE TABLE IF NOT EXISTS Names" + postfix + " (id INTEGER PRIMARY KEY NOT NULL , Name TEXT NOT NULL COLLATE NOCASE )",
             });
         }
